Rename `db` module to `providers` to avoid name conflict.
