--- virtool_workflow/abc/db/data_provider.py
@@ -1,10 +1,0 @@
-from abc import ABC
-from .analysis import AbstractAnalysisProvider
-from .cache import AbstractCacheProvider
-
-
-class AbstractDataProvider(ABC):
-    analysis: AbstractAnalysisProvider
-    caches: AbstractCacheProvider
-
-
