+++ virtool_workflow/abc/providers/references.py
@@ -1,4 +1,4 @@
-from abc import ABC, abstractmethod
+from abc import ABC
 
 
 class AbstractReferenceProvider(ABC):
